py_requirements.txt
numpy~=1.22.4
pandas~=1.4.2
setuptools~=61.2.0
scipy~=1.7.3
matplotlib~=3.5.1
requests~=2.27.1
tqdm~=4.64.0
biopython